pyproject.toml
[tool.poetry]
name = "beer-full-stack"
version = "0.1.0"
description = ""
authors = ["Corey DeJong <dejongcorey@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.8.5"
django = "^3.0.8"
djangorestframework = "^3.11.0"
gunicorn = "^20.0.4"
whitenoise = "^5.1.0"
django-cors-headers = "^3.4.0"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
